Basic start of the website.
This is something that could be really worked on in the future as there's way to many breaks and things inside of the actual code. But it get's the basic idea out there.

Also need to add things such as the actual mobile navigation along with other pages (other pages are easier once you get the base layers down).

I also need to add a footer allowing me to be able to add more information at the bottom of the page.